# Use sync.Pool to use &item{}