```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkConcurrentCacheSetNeverEvict-12     96.3          127           +31.88%

benchmark                                    old allocs     new allocs     delta
BenchmarkConcurrentCacheSetNeverEvict-12     3              2              -33.33%

benchmark                                    old bytes     new bytes     delta
BenchmarkConcurrentCacheSetNeverEvict-12     110           108           -1.82%
```
<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/ristretto/136)
<!-- Reviewable:end -->