@CLAassistant 2020-04-29
[![CLA assistant check](https://cla-assistant.io/pull/badge/signed)](https://cla-assistant.io/dgraph-io/ristretto?pullRequest=136) <br/>All committers have signed the CLA.